+++ CHANGELOG.md
@@ -5,7 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-[Unreleased](https://github.com/jshwi/pyaud/compare/v1.1.1...HEAD)
+[Unreleased](https://github.com/jshwi/pyaud/compare/v1.2.0...HEAD)
+------------------------------------------------------------------------
+
+[1.2.0](https://github.com/jshwi/pyaud/releases/tag/v1.2.0) - 2021-02-06
 ------------------------------------------------------------------------
 ### Added
 - Adds ``L0G_LEVEL`` environment variable to permanently set default loglevel to other from ``INFO``